requirements.txt
accelerate==1.11.0
aiodns==3.5.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
aistudio-sdk==0.3.8
altair==5.5.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==3.7.1
attrdict==2.0.1
attrs==25.4.0
azure-core==1.36.0
azure-identity==1.25.1
babel==2.17.0
backoff==2.2.1
backports.zstd==1.0.0
bce-python-sdk==0.9.46
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.56
botocore==1.40.56
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
comtypes==1.4.13
contourpy==1.3.3
cryptography==45.0.7
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.5
dashscope==1.24.7
defusedxml==0.7.1
dill==0.4.0
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
dotenv==0.9.9
easyocr==1.7.2
einops==0.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.119.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
ffmpy==0.6.4
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
groq==0.33.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
identify==2.6.15
idna==3.11
imageio==2.37.0
imagesize==1.4.1
imgaug==0.4.0
iniconfig==2.3.0
inquirerpy==0.3.4
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
modelscope==1.31.0
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
narwhals==2.9.0
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==1.3.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.3
packaging==25.0
paddleocr==2.7.3
paddlepaddle==3.2.0
paddlex==3.3.4
pandas==2.3.3
paramiko==4.0.0
pdf2docx==0.5.8
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
pre-commit==3.8.0
premailer==3.10.0
prettytable==3.16.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycares==4.11.0
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.10
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==2.0.1
PyMuPDF==1.26.5
PyNaCl==1.6.0
pyparsing==3.2.5
pypdfium2==4.30.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.3.3
pytest-asyncio==0.23.6
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
rarfile==4.2
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.14.1
runpod==1.7.13
s3transfer==0.14.0
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
screeninfo==0.8.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.42.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
streamlit==1.50.0
supervision==0.18.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
tifffile==2025.10.16
timm==1.0.20
tokenizers==0.21.4
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
tqdm-loggable==0.2
transformers==4.51.2
triton==3.5.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uiautomation==2.0.29
ujson==5.11.0
ultralytics==8.3.70
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.3
visualdl==2.5.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==14.2
Werkzeug==3.1.3
yarl==1.22.0
